Move commit_tasks_status, document the documentId path for each Firestore model.
