+++ app_dart/test/request_handlers/get_green_commits_test.dart
@@ -7,7 +7,7 @@ import 'dart:convert';
 import 'package:cocoon_server_test/test_logging.dart';
 import 'package:cocoon_service/cocoon_service.dart';
 import 'package:cocoon_service/src/model/appengine/task.dart';
-import 'package:cocoon_service/src/model/firestore/commit_tasks_status.dart';
+import 'package:cocoon_service/src/service/build_status_provider/commit_tasks_status.dart';
 import 'package:test/test.dart';
 
 import '../src/datastore/fake_config.dart';
