+++ app_dart/test/service/commit_tasks_status_test.dart
@@ -3,7 +3,7 @@
 // found in the LICENSE file.
 
 import 'package:cocoon_server_test/test_logging.dart';
-import 'package:cocoon_service/src/model/firestore/commit_tasks_status.dart';
+import 'package:cocoon_service/src/service/build_status_provider/commit_tasks_status.dart';
 import 'package:test/test.dart';
 
 import '../../src/utilities/entity_generators.dart';
