+++ app_dart/test/src/service/fake_build_status_provider.dart
@@ -2,8 +2,8 @@
 // Use of this source code is governed by a BSD-style license that can be
 // found in the LICENSE file.
 
-import 'package:cocoon_service/src/model/firestore/commit_tasks_status.dart';
 import 'package:cocoon_service/src/service/build_status_provider.dart';
+import 'package:cocoon_service/src/service/build_status_provider/commit_tasks_status.dart';
 import 'package:collection/collection.dart';
 import 'package:github/github.dart';
